refactor polymer.py (third draft)
