+++ meeko/residue_chem_templates.py
@@ -246,4 +246,3 @@ class ResidueChemTemplates:
         if len(missing):
             raise ValueError(f"missing residue templates for ambiguous: {missing}")
         return
-
